tests/index.html

<html>
<head>
		<script type="text/javascript">
			function print(to_print)
			{
				document.write(to_print + '<br>');
			}
		</script>
    <script type="text/javascript" src="../mochikit/MochiKit/MochiKit.js"></script>
    <script type="text/javascript" src="../mochikit/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="../MochiChi/MochiChi.js"></script>
    <script type="text/javascript" src="test_Parsing.js"></script>
    <link rel="stylesheet" type="text/css" href="../mochikit/tests/SimpleTest/test.css">
    <!--<script type="text/javascript" src="test_Connection.js"></script>-->
</head>
<body>
<pre id="test">
<script type="text/javascript" src="test_Base.js"></script>
<script type="text/javascript">
try {
    SimpleTest.saveScope(this);
    tests.test_Parsing({ok:ok,is:is});
    SimpleTest.verifyScope(this);
    ok( true, "test suite finished!");
} catch (err) {
    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );
}
</script>
</pre>
</body>
</html>
